Add support for transform_function_options
For transform functions to be reusable, it may be useful to be able
to pass additional parameters other than the inventory host object.
A typical use case is to pass in sensitive configuration data from
a separate secret management tool to the inventory data.

This change adds an additional `transform_function_options` parameter
to the `InventoryConfig`.

The transform_function_options` must be a dictionary with string keys
which are also valid python identifiers. The contents of this dict
will be passed to the transform function as `kwargs`.

Example:

```yaml
---
inventory:
  transform_function: __main__.transform_inventory
  transform_function_options:
    password: "supersecret"
```

```python
def transform_inventory(host, password, **kwargs):
  """ Example transform function which adds the password argument to
  each host data
  """
  host.data['password'] = password
```

Add docs and tests for transform_function_options

Fixes 290