Remove the external ping test (#293)
* Remove the external ping test

By having a test for external connectivity which isn't stubbed,
what's actually being tested is the network environment of the development
machine, and not the nornir code. There are several examples where
this test would legitimately fail:
- Offline laptop
- Corporate environment with restrictive firewall policies.

As such, a failing test does not necessarily indicate an issue with the
code. A failed test will cause downstream tools that depend on passing
tests to fail, or leave developers wasting time looking for non-existent
faults.

* Remove resources used by external ping test