Add support for deep merging config file, env and kwargs settings
This commit makes it possible to have non-overlapping and overlapping
settings present in all of configuration file, environment variables
and passed to `InitNornir` as kwargs, with all three merged with
respect to the precedence order of:
`hardcoded defaults < config file < env vars < kwargs`.

Previous to this commit:
- Env vars would only be considered if the setting was not present
  in the config file at all
- If any kwarg was set, providing a top-level category of configuration
  (e.g. `core`, `inventory`, `logging`, etc) then that kwarg would be
  taken exclusively, and any env vars or config file settings would
  be disregarded.

To maintain backwards compatibility with the 2.0 behaviour, this new
functionality is opt-in by passing `deep_merge=True` to `InitNornir`
or to the `Config.load_from_file` functions. The default behaviour is
to use the kwarg exclusively if provided, and take the merger of
config file and env var settings otherwise.

Includes tests and documentation.

Fixes for #291