Add support for deep merging config file, env and kwargs settings
This commit makes it possible to have non-overlapping and overlapping
settings present in all of configuration file, environment variables
and passed to `InitNornir` as kwargs, with all three merged fully
recursively, with respect to the precedence order of:
`hardcoded defaults < config file < env vars < kwargs`.

Previous to this commit, only immediate children of each top-level
settings section would be merged.

To maintain backwards compatibility with the 2.0 behaviour, this new
functionality is opt-in by passing `deep_merge=True` to `InitNornir`
or to the `Config.load_from_file` functions. The default behaviour is
to shallow merge the immediate children of each top-level settings
section.

Includes tests and documentation.

Fixes for #291