Update confusing tutorial ordering
Initializing Nornir before creating an inventory file throws a FileNotFoundError for the missing inventory/hosts.yaml file. This was unexpected as a new user moving through the documentation linearly. An easy solution is to introduce the reader to the inventory section before the initialization section. Otherwise, a note on expecting the error would suffice.